GEOS-8666: wpsResourceManager cleanup - fix - 2.13.x bkp

+++ src/extension/wps/wps-core/src/main/java/org/geoserver/wps/resource/WPSResourceManager.java
@@ -427,11 +427,12 @@ public class WPSResourceManager extends ProcessListenerAdapter implements Dispat
         if (resourceType == Type.RESOURCE && resource.lastmodified() < expirationThreshold) {
             result = resource.delete();
         } else if (resourceType == Type.DIRECTORY) {
+            long directoryModified = resource.lastmodified();
             for (Resource child : resource.list()) {
                 result &= cleanupResource(child, expirationThreshold);
             }
             // Cleanup the directory too if all the children have been cleanup
-            if (result) {
+            if (result && directoryModified < expirationThreshold) {
                 result &= resource.delete();
             }
         }
